requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3
gunicorn==21.2.0
